test(web): split locations subtab tests
